requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.25
httpx==0.26.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0
aiofiles==23.2.1
aiosqlite==0.19.0
